+++ src/assets/Fonts.tsx
@@ -21,16 +21,37 @@ const Fonts = () => (
             @font-face {
                 font-family: 'Gattica';
                 font-style: normal;
-                font-weight: 500;
-                font-dispaly: swap;
-                src: url("/fonts/Gattica-Medium100.otf") format("opentype");
-            }
-            @font-face {
-                font-family: 'Gattica';
-                font-style: normal;
                 font-weight: 700;
                 font-dispaly: swap;
                 src: url("/fonts/Gattica-Bold100.otf") format("opentype");
+            }
+            @font-face {
+                font-family: 'Inter';
+                font-style:  normal;
+                font-weight: 400;
+                font-display: swap;
+                src: url("/fonts/Inter-Regular.ttf") format("truetype");
+            }
+            @font-face {
+                font-family: 'Inter';
+                font-style:  normal;
+                font-weight: 500;
+                font-display: swap;
+                src: url("/fonts/Inter-Medium.ttf") format("truetype");
+            }
+            @font-face {
+                font-family: 'Inter';
+                font-style:  normal;
+                font-weight: 600;
+                font-display: swap;
+                src: url("/fonts/Inter-SemiBold.ttf") format("truetype");
+            }
+            @font-face {
+                font-family: 'Inter';
+                font-style:  normal;
+                font-weight: 700;
+                font-display: swap;
+                src: url("/fonts/Inter-Bold.ttf") format("truetype");
             }
                  `}
     />